refactor(backend): swap raw Bun.serve for Hono framework (#57 step 3)
Migrate HTTP layer from raw Bun.serve + 953-line pathname-matching
switch in app.ts to Hono framework with route modules.

Why: type-safe routing via radix tree, declarative zod validation
middleware (kills parseJson helper), built-in cookie + cors helpers
(drop `cookie` + `@types/cookie` deps), splits app.ts into focused
route files (resolves §3 file-size violation flagged in #74 review).

Structure:
- backend/src/api/index.ts: Hono app composition, middleware order,
  route mounting, public-auth-path allowlist
- backend/src/api/middleware.ts: securityHeaders, originGuard (custom
  CORS preserving 403 ORIGIN_NOT_ALLOWED envelope), sessionGuard
- backend/src/api/responses.ts: jsonData/jsonError helpers + validateJson
  wrapper around @hono/zod-validator preserving VALIDATION_ERROR shape
- backend/src/api/session.ts: createSession/deleteSession/cookies/bcrypt
- backend/src/api/{auth,transactions,movements,snapshots,styles,prefs,
  backup}.ts: route modules per resource
- backend/src/api/backup-helpers.ts: shared buildBackupPayload,
  wipeUserData, applyImport (extracted from app.ts closures)
- backend/src/security-headers.ts: single source of CSP + other headers,
  shared between Hono middleware and server.ts static-file handler
- backend/src/app.ts: shrunk from 953 lines to 2-line re-export for
  backward compat with test-helpers

API contract preserved (frontend untouched):
- Response envelopes byte-identical: {data} success, {error: {code,
  message, fields?}} on failure
- All 5 security headers attached to EVERY response (including 403
  ORIGIN_NOT_ALLOWED and xlsx downloads)
- Session cookie attributes unchanged (HttpOnly, SameSite=Strict,
  path=/, maxAge, conditional Secure via COOKIE_SECURE env)
- bcrypt → argon2id rehash on login preserved
- Disabled-account guard enforced at login AND sessionGuard
- 401 for unauth, 404 for authed-but-unknown route, 405 for non-API
  non-GET — all preserved
- XLSX upload: 10 MB cap, MIME validation, multipart + base64 paths

Removed:
- 953-line createApi closure in app.ts
- helpers.ts: setSecurityHeaders, makeError, makeData, parseJson
  (replaced by Hono machinery; pure helpers inferType/makeId/
  normalizeTx|Mm|Snap kept)
- cookie + @types/cookie deps (hono/cookie replaces)

Added: hono ^4.12.23, @hono/zod-validator ^0.8.0

Verification:
- bunx tsc --noEmit: clean
- bun test: 125/125 pass (zero modifications to existing test files
  needed — strongest evidence API contract preserved)
- npm run typecheck (backend tsc + frontend build): clean
- Pre-PR Opus reviewer: verdict pass, 0 critical, 0 AGENTS violations,
  0 security concerns. 2 suggestions addressed (DRY SECURITY_HEADERS
  via shared module, simplified CORS double-set in originGuard).
