+++ backend/package.json
@@ -10,16 +10,16 @@
     "test": "../scripts/run-bun.sh test --pass-with-no-tests"
   },
   "dependencies": {
+    "@hono/zod-validator": "^0.8.0",
     "bcryptjs": "^3.0.3",
-    "cookie": "^1.1.1",
     "drizzle-orm": "^0.45.2",
+    "hono": "^4.12.23",
     "xlsx": "^0.18.5",
     "zod": "^4.4.3"
   },
   "devDependencies": {
     "@types/bcryptjs": "^3.0.0",
     "@types/bun": "^1.3.14",
-    "@types/cookie": "^1.0.0",
     "@types/node": "^25.9.1",
     "drizzle-kit": "^0.31.10",
     "typescript": "^6.0.3"
