+++ backend/src/api/styles.ts
@@ -1,0 +1,53 @@
+import { Hono } from "hono";
+
+import { eq } from "drizzle-orm";
+import { getDrizzle } from "../db.ts";
+import { asset_styles } from "../db/schema.ts";
+import { stylesSchema } from "../schemas.ts";
+import type { AppEnv } from "./types.ts";
+import { jsonData, validateJson } from "./responses.ts";
+
+export const stylesRoutes = new Hono<AppEnv>();
+
+stylesRoutes.get("/", (c) => {
+  const db = c.get("db");
+  const user = c.get("user");
+  const rows = getDrizzle(db)
+    .select({
+      asset: asset_styles.asset,
+      color_hex: asset_styles.color_hex,
+      risk_level: asset_styles.risk_level
+    })
+    .from(asset_styles)
+    .where(eq(asset_styles.user_id, user.id))
+    .all();
+  const styles: Record<string, { colorHex: string | null; riskLevel: string | null }> = {};
+  rows.forEach((row) => {
+    styles[row.asset] = { colorHex: row.color_hex ?? null, riskLevel: row.risk_level ?? null };
+  });
+  return jsonData(c, styles);
+});
+
+stylesRoutes.put("/", validateJson(stylesSchema), (c) => {
+  const body = c.req.valid("json");
+  const db = c.get("db");
+  const user = c.get("user");
+  const dbo = getDrizzle(db);
+  const tx = db.transaction(() => {
+    dbo.delete(asset_styles).where(eq(asset_styles.user_id, user.id)).run();
+    for (const [asset, style] of Object.entries(body.styles)) {
+      if (!asset.trim()) continue;
+      dbo
+        .insert(asset_styles)
+        .values({
+          user_id: user.id,
+          asset: asset.trim(),
+          color_hex: style.colorHex ?? null,
+          risk_level: style.riskLevel ?? null
+        })
+        .run();
+    }
+  });
+  tx();
+  return jsonData(c, { ok: true });
+});
