+++ backend/src/helpers.ts
@@ -1,45 +1,3 @@
-import type { z } from "zod";
-
-export function setSecurityHeaders(h: Headers): void {
-  h.set("x-content-type-options", "nosniff");
-  h.set("x-frame-options", "DENY");
-  h.set("referrer-policy", "no-referrer");
-  h.set("strict-transport-security", "max-age=63072000; includeSubDomains");
-  h.set(
-    "content-security-policy",
-    "default-src 'self'; img-src 'self' data:; style-src 'self' 'unsafe-inline'; script-src 'self'; connect-src 'self'; frame-ancestors 'none'; base-uri 'self'; form-action 'self'"
-  );
-}
-
-export function makeError(
-  code: string,
-  message: string,
-  status = 400,
-  fields?: Record<string, string>,
-  headers?: Headers
-): Response {
-  const h = headers ? new Headers(headers) : new Headers();
-  h.set("content-type", "application/json");
-  setSecurityHeaders(h);
-  return new Response(JSON.stringify({ error: { code, message, fields } }), { status, headers: h });
-}
-
-export function makeData(data: unknown, status = 200, headers?: Headers): Response {
-  const h = headers ? new Headers(headers) : new Headers();
-  h.set("content-type", "application/json");
-  setSecurityHeaders(h);
-  return new Response(JSON.stringify({ data }), { status, headers: h });
-}
-
-export async function parseJson<T>(req: Request, schema: z.ZodType<T>): Promise<T> {
-  const raw = await req.json().catch(() => null);
-  const parsed = schema.safeParse(raw);
-  if (!parsed.success) {
-    throw makeError("VALIDATION_ERROR", "Invalid request body", 400);
-  }
-  return parsed.data;
-}
-
 export function makeId(prefix: string): string {
   return `${prefix}-${crypto.randomUUID()}`;
 }
